Growth rate estimate using MCMC

Uses Rstan and the No U-turn sampler to approximate the growth rate using the likelihood from Stadler 2009 "On incomplete sampling under birth–death models and connections to the sampling-based coalescent"

Usage

birthDeathMCMC(
  tree,
  maxGrowthRate = 4,
  alpha = 0.05,
  verbose = TRUE,
  nChains = 3,
  nCores = 1,
  chainLength = 2000
)

Arguments

tree

An ultrametric tree subset to include only the clone of interest. Alternatively, a list with several such trees.

maxGrowthRate

Sets upper bound on birth rate. Default is 4 but this will depend on the nature of the data

alpha

Used for calculation of confidence intervals. 1-alpha confidence intervals used with default of alpha = 0.05 (95 percent confidence intervals)

verbose

TRUE or FALSE, should the Rstan MCMC intermediate output and progress be printed?

nChains

Number of chains to run in MCMC. Default is 3

nCores

Number of cores to perform MCMC. Default is 1, but chains can be run in parallel

chainLength

Number of iterations for each chain in MCMC. Default is 2000, increase if stan tells you to

Value

A dataframe including the net growth rate estimate, confidence intervals, and other important details (clone age estimate, runtime, n, etc.)

See also

Examples

df <- birthDeathMCMC(cloneRate::exampleUltraTrees[[1]])
#> 
#> CHECKING DATA AND PREPROCESSING FOR MODEL '074955cb23b8ac2bef05bf679ec41d0b' NOW.
#> 
#> COMPILING MODEL '074955cb23b8ac2bef05bf679ec41d0b' NOW.
#> 
#> STARTING SAMPLER FOR MODEL '074955cb23b8ac2bef05bf679ec41d0b' NOW.
#> 
#> SAMPLING FOR MODEL '074955cb23b8ac2bef05bf679ec41d0b' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000127 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.27 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 6.014 seconds (Warm-up)
#> Chain 1:                6.68325 seconds (Sampling)
#> Chain 1:                12.6972 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL '074955cb23b8ac2bef05bf679ec41d0b' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 5.6e-05 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 0.56 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 6.41143 seconds (Warm-up)
#> Chain 2:                6.96159 seconds (Sampling)
#> Chain 2:                13.373 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL '074955cb23b8ac2bef05bf679ec41d0b' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 5.5e-05 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 0.55 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 6.38784 seconds (Warm-up)
#> Chain 3:                6.83914 seconds (Sampling)
#> Chain 3:                13.227 seconds (Total)
#> Chain 3: